deployment/requirements.txt
requests
fastapi
uvicorn[standard]
opencv-python
python-multipart
ndjson
google.cloud
pytorch-lightning
timm
httpx
fastapi-scheduler
aiosqlite
Pillow
transformers
numpy
pandas
evidently
